Moved general user commands out of the cog folder
